speech-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
numpy<2.0.0
openai-whisper==20231117
torch==2.2.1
python-dotenv==1.0.1
